requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
dj-places==5.1.0
Django==3.2.15
googlemaps==4.6.0
idna==3.3
phonenumbers==8.12.54
psycopg2==2.9.3
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.12
